@@ -5,27 +5,41 @@
   "private": true,
   "main": "js/bundle.js",
   "scripts": {
-    "start": "beefy js/index.js:js/bundle.js --live --open -- -t reactify",
+    "start": "watchify -v -e js/index.js -o js/bundle.js",
+    "build": "browserify -e js/index.js -o js/bundle.js",
+    "watch": "watchify -v -e js/index.js -o js/bundle.js",
+    "beefy": "beefy js/index.js:js/bundle.js --live --open",
+    "serve": "live-server --watch=js/bundle.js,index.html",
     "less": "lessc styles/app.less > styles/app.css",
-    "prod": "browserify -e js/index.js -o js/bundle.js -t reactify -g uglifyify",
+    "prod": "browserify -e js/index.js -o js/bundle.js -g uglifyify",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Dheeraj V S",
   "license": "Proprietary",
+  "browserify": {
+    "transform": [
+      "babelify"
+    ]
+  },
   "dependencies": {
-    "bootstrap": "^3.3.5",
-    "lodash": "^3.10.0",
-    "material-ui": "^0.10.1",
+    "bootstrap": "^3.3.6",
+    "lodash": "^4.11.1",
+    "material-ui": "^0.15.0-beta.2",
     "numeral": "^1.5.3",
-    "react": "^0.13.3",
-    "react-bootstrap": "^0.23.7",
-    "react-tap-event-plugin": "^0.1.7",
-    "superagent": "^1.2.0"
+    "react": "^15.0.2",
+    "react-dom": "^15.0.2",
+    "react-bootstrap": "^0.29.3",
+    "react-router": "^2.4.0",
+    "react-tap-event-plugin": "^1.0.0",
+    "superagent": "^1.8.3"
   },
   "devDependencies": {
-    "browserify": "^10.2.6",
-    "react-router": "^0.13.3",
-    "reactify": "^1.1.1",
-    "uglifyify": "^3.0.1"
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "babelify": "^7.3.0",
+    "browserify": "^13.0.0",
+    "uglifyify": "^3.0.1",
+    "watchify": "^3.7.0"
   }
 }
