@@ -2,26 +2,22 @@
  * Copyright (c) 2015, Sowmya B A. All rights reserved.
  */
 
-'use strict';
+ import React from 'react';
+ import ReactDOM from 'react-dom';
+ import injectTapEventPlugin from 'react-tap-event-plugin';
+ import Routes from './components/Routes';
 
-var React = require('react'),
-  Router = require('react-router'),
-  routes = require('./routes'),
-  injectTapEventPlugin = require('react-tap-event-plugin');
+ // Needed for onTouchTap; Can go away when react 1.0 release
+ // Check this repo: https://github.com/zilverline/react-tap-event-plugin
+ injectTapEventPlugin();
 
-  //Needed for onTouchTap
-  //Can go away when react 1.0 release
-  //Check this repo:
-  //https://github.com/zilverline/react-tap-event-plugin
-  injectTapEventPlugin();
+ const appElement = document.getElementById('app');
 
-Router.run(routes, function (Root) {
-  React.render(<Root/>, document.body);
-});
+ ReactDOM.render(Routes, appElement)
 
-// Prevent context menu on img elements to prevent saving images
-document.body.addEventListener('contextmenu', function (e) {
-  if (e.srcElement.nodeName === 'IMG') {
-    e.preventDefault();
-  }
-});
+ // Prevent context menu on img elements to prevent saving images
+ appElement.addEventListener('contextmenu', function (e) {
+   if (e.srcElement.nodeName === 'IMG') {
+     e.preventDefault();
+   }
+ });
